+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.3.2+1
+
+* Support `package:analyzer` `0.39.0`.
+
 # 1.3.2
 
 * Restore the code that publishes the dart-style npm package.